refactor: Replace navigation group routes with conditional rendering
