--- frontend/src/app/[locale]/(navigation)/layout.module.scss
@@ -1,35 +1,0 @@
-@use '@/styles/grid-system' as gs;
-@use '@/styles/breakpoints' as bp;
-
-.navigation-layout {
-  box-sizing: border-box;
-  position: relative;
-  width: 100%;
-  height: 100%;
-  overflow-y: auto;
-
-  @include gs.grid-system;
-
-  @include bp.respond-to('compact') {
-    padding: 80px 16px 96px;
-    // background-color: var(--md-sys-color-background);
-  }
-
-  @include bp.respond-to('medium') {
-    padding: 24px 32px;
-    // background-color: var(--md-sys-color-background);
-  }
-
-  @include bp.respond-to('expanded') {
-    padding: 36px 24px;
-    // background-color: var(--md-sys-color-background);
-  }
-
-  @include bp.respond-to('large') {
-    padding: 64px 200px;
-  }
-
-  @include bp.respond-to('extra-large') {
-    padding: 96px 236px;
-  }
-}
