--- frontend/src/app/[locale]/(no-navigation)/login/login.module.scss
@@ -1,16 +1,0 @@
-.page {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  height: 100%;
-
-  gap: 150px;
-}
-
-.login-button {
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-}
