--- frontend/src/app/[locale]/(no-navigation)/login/page.tsx
@@ -1,14 +1,0 @@
-import Image from 'next/image';
-
-import styles from './login.module.scss';
-
-export default function LoginPage() {
-  return (
-    <div className={styles['page']}>
-      <Image src="/logo.svg" width={267} height={223} alt="logo" />
-      <button className={styles['login-button']}>
-        <Image src="/login.svg" width={173} height={40} alt="google login" />
-      </button>
-    </div>
-  );
-}
