+++ frontend/src/app/[locale]/difficulty/[level]/page.tsx
