+++ frontend/src/app/[locale]/difficulty/page.tsx
