+++ frontend/src/app/[locale]/layout.module.scss
@@ -1,16 +1,18 @@
 @use '@/styles/grid-system' as gs;
 @use '@/styles/breakpoints' as bp;
 
-.layout-container-main {
+.navigation-layout {
   box-sizing: border-box;
+  position: relative;
   width: 100%;
   height: 100%;
   overflow-y: auto;
+  background-color: var(--md-sys-color-background);
 
   @include gs.grid-system;
 
   @include bp.respond-to('compact') {
-    padding: 16px;
+    padding: 80px 16px 96px;
   }
 
   @include bp.respond-to('medium') {
