+++ frontend/src/app/[locale]/layout.tsx
@@ -1,5 +1,17 @@
+import Navigation from '@/components/Navigation/Navigation';
 import I18nProvider from '../../providers/I18nProvider';
 
-export default async function LocaleLayout({ children }: { children: React.ReactNode }) {
-  return <I18nProvider>{children}</I18nProvider>;
-}
+import styles from './layout.module.scss';
+
+const NavigationLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <div className={styles['navigation-layout']}>
+      <I18nProvider>
+        <Navigation />
+        {children}
+      </I18nProvider>
+    </div>
+  );
+};
+
+export default NavigationLayout;
