+++ frontend/src/app/[locale]/learning/layout.module.scss
