+++ frontend/src/app/[locale]/learning/page.tsx
