+++ frontend/src/app/[locale]/login/page.tsx
@@ -1,0 +1,16 @@
+import Image from 'next/image';
+
+import styles from './login.module.scss';
+
+export default function LoginPage() {
+  return (
+    <div className={styles['page']}>
+      <div className={styles['container']}>
+        <Image src="/logo.svg" width={267} height={223} alt="logo" />
+        <button className={styles['login-button']}>
+          <Image src="/login.svg" width={173} height={40} alt="google login" />
+        </button>
+      </div>
+    </div>
+  );
+}
