+++ frontend/src/app/[locale]/meanings/[category]/page.tsx
