+++ frontend/src/app/[locale]/meanings/page.tsx
