+++ frontend/src/app/[locale]/settings/Settings.module.scss
