+++ frontend/src/app/[locale]/settings/page.tsx
